explorer/charts: fix int32 overflow risk and add zero-canvas guard
- Upgrade TotalBlocksMined return type from int32 to int64 in DataSource
  interface, ChainDB, and API response to prevent silent overflow on long-
  lived chains. SQL casts updated from INT4 to INT8.

- Add requestAnimationFrame guard in renderHashrateShares when
  getBoundingClientRect returns zero width/height (edge case where
  container toggling hasn't propagated layout yet)

+++ cmd/dcrdata/internal/api/apiroutes.go
@@ -112,7 +112,7 @@ type DataSource interface {
 	LoadSKASupplyForCoin(ctx context.Context, charts *cache.ChartData, coinType uint8) error
 	LoadSKAFeesForCoin(ctx context.Context, charts *cache.ChartData, coinType uint8) error
 	GetMinerHashrateShares(ctx context.Context, since *time.Time) ([]dbtypes.MinerShareData, error)
-	GetTotalBlocksMined(ctx context.Context, since *time.Time) (int32, error)
+	GetTotalBlocksMined(ctx context.Context, since *time.Time) (int64, error)
 }
 
 // dcrdata application context used by all route handlers
@@ -1976,7 +1976,7 @@ func (c *appContext) ChartTypeData(w http.ResponseWriter, r *http.Request) {
 
 		resp := struct {
 			Interval string                   `json:"interval"`
-			Total    int32                    `json:"total"`
+			Total    int64                    `json:"total"`
 			Miners   []dbtypes.MinerShareData `json:"miners"`
 		}{
 			Interval: interval,
